refactor : Verification 클래스 성격을 고려해 이름을 InputConverter로 변경
